requirements.txt
absl-py==2.3.1
accelerate==1.11.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
apex==0.1
async-timeout==5.0.1
attrs==25.4.0
av==16.0.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
bert-score==0.3.13
bitsandbytes==0.48.1
boto3==1.42.2
botocore==1.42.2
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contexttimer==0.3.3
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
datasets==2.19.0
debugpy==1.8.17
decorator==5.2.1
decord==0.6.0
Deprecated==1.2.18
diffusers @ git+https://github.com/huggingface/diffusers.git@84e16575e4c5e90b6b49301cfa162ced4cf478d2
dill==0.3.8
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
easydict==1.13
einops==0.8.1
email-validator==2.3.0
entmax==1.3
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fabric==3.2.2
fastapi==0.115.14
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
ffmpy==0.6.4
filelock==3.19.1
flash_attn @ file:///mnt/afs/250010063/Codes/SurGen/flash_attn-2.8.3%2Bcu12torch2.4cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=f6e2c4172c98a8225b45424dd87ab43224df878cfb4148d679e2c98a7a7dd24d
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2024.3.1
ftfy==6.3.1
galore-torch==1.0
gitdb==4.0.12
GitPython==3.1.45
google==3.0.0
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
hanlp==2.1.3
hanlp_common==0.0.23
hanlp_downloader==0.0.25
hanlp_trie==0.0.5
hatch-vcs==0.5.0
hatchling==1.27.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
identify==2.6.15
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
invoke==2.2.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
jsonargparse==3.13.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
language_tool_python==3.1.0
lightning-utilities==0.15.2
loguru==0.7.3
lpips==0.1.4
lxml==6.0.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mmengine==0.10.7
modelscope==1.33.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
networkx==3.3
ninja==1.13.0
nltk==3.9.2
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==13.590.44
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.109.1
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
orjson==3.11.4
packaging==25.0
pandas==2.3.3
paramiko==4.0.0
pathspec==0.12.1
peft==0.18.0
phrasetree==0.0.9
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
plumbum==1.9.0
portalocker==3.2.0
pre_commit==4.3.0
propcache==0.4.1
protobuf==4.25.8
psutil==7.1.2
pyarrow==22.0.0
pyarrow-hotfix==0.7
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.8.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyNaCl==1.6.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-slugify==8.0.4
pytorch-fid==0.3.0
pytorch-lightning==2.6.0
pytz==2025.2
PyYAML==6.0.3
ray==2.50.1
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
rpds-py==0.28.0
rpyc==6.0.0
ruff==0.14.2
s3transfer==0.16.0
sacrebleu==2.5.1
safehttpx==0.1.7
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.42.1
setuptools-scm==9.2.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
text-unidecode==1.3
threadpoolctl==3.6.0
timm==1.0.21
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
tomlkit==0.13.3
toposort==1.5
torch==2.6.0+cu124
torch-dct==0.1.6
torch-fidelity==0.3.0
torchaudio==2.6.0+cu124
TorchCodec==0.1.0
torcheval==0.0.7
torchmetrics==0.10.3
torchvision==0.21.0+cu124
tqdm==4.67.1
transformers==4.57.1
triton==3.2.0
trove-classifiers==2025.9.11.17
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unbabel-comet==2.2.7
urllib3==2.5.0
uvicorn==0.29.0
uvloop==0.22.1
virtualenv==20.35.3
wandb==0.22.3
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
zipp==3.23.0
